# swagger-cloudtalkapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CloudTalk API",
...
    "contact": {
      "name": "CloudTalk API Support",
      "url": "https://www.cloudtalk.io/contact"
    },
    "version": "1.7",
    "x-logo": {
      "url": "https://my.cloudtalk.io/img/logo-cloudtalk-color.png"
    }
  },
  "servers": [
    {
      "url": "https://my.cloudtalk.io/api"
    }
  ],
...
  "paths": {
...
    "/calls/{callId}": {
      "get": {
        "servers": [
          {
            "url": "https://analytics-api.cloudtalk.io/api/"
          }
        ],
        "tags": ["Calls"],
        "summary": "Comprehensive information about a call and its flow",
        "parameters": [
          {
            "name": "callId",
            "in": "path",
            "required": true,
            "description": "ID of a call",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about a given call",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Call"
                },
                "example": {
                  "cdr_id": 12345,
                  "uuid": "aaaa-bbbb-cccc-dddd",
                  "company_id": 100123,
                  "date": "2022-02-07T13:00:00.000Z",
                  "contact": {
                    "id": 1,
                    "name": "John Doe",
                    "country": "CZ",
                    "number": "+420123123123"
                  },
                  "call_tags": [
                    {
                      "id": 5,
                      "label": "Lead"
                    }
                  ],
                  "call_rating": 3,
                  "internal_number": {
                    "id": 1,
                    "name": "Europe Sales",
                    "number": "+421456456456"
                  },
                  "call_times": {
                    "talking_time": 60,
                    "wrap_up_time": 10,
                    "ringing_time": 10,
                    "total_time": 80,
                    "waiting_time": 20,
                    "holding_time": 5
                  },
                  "direction": "incoming",
                  "type": "regular",
                  "status": "answered",
                  "call_steps": [
                    {
                      "type": "ivr",
                      "id": 1,
                      "date": "2022-02-07T13:00:00.000Z",
                      "total_time": 10,
                      "option": "1"
                    },
                    {
                      "type": "queue",
                      "id": 10,
                      "name": "Queue10",
                      "date": "2022-02-07T13:00:10.000Z",
                      "call_times": {
                        "talking_time": 60,
                        "wrap_up_time": 10,
                        "ringing_time": 10,
                        "total_time": 80,
                        "waiting_time": 10,
                        "holding_time": 3
                      },
                      "status": "answered",
                      "reason": null,
                      "resolved_by": null,
                      "resolved_by_call": {
                        "id": null,
                        "date": null,
                        "agent": {
                          "id": null,
                          "name": null
                        }
                      },
                      "strategy": "rrmemory",
                      "agent_calls": [
                        {
                          "type": "agent",
                          "id": 12,
                          "name": "Agent12",
                          "group_ids": [10, 20],
                          "date": "2022-02-07T13:00:10.000Z",
                          "call_times": {
                            "talking_time": null,
                            "wrap_up_time": null,
                            "ringing_time": 5,
                            "total_time": 5,
                            "holding_time": null
                          },
                          "status": "missed",
                          "reason": "offline",
                          "resolved_by": null,
                          "resolved_by_call": {
                            "id": null,
                            "date": null,
                            "agent": {
                              "id": null,
                              "name": null
                            }
                          }
                        },
                        {
                          "type": "agent",
                          "id": 11,
                          "name": "Agent11",
                          "group_ids": [10, 20],
                          "date": "2022-02-07T13:00:15.000Z",
                          "call_times": {
                            "talking_time": 60,
                            "wrap_up_time": 10,
                            "ringing_time": 5,
                            "total_time": 75,
                            "holding_time": 2
                          },
                          "status": "answered",
                          "reason": null,
                          "resolved_by": null,
                          "resolved_by_call": {
                            "id": null,
                            "date": null,
                            "agent": {
                              "id": null,
                              "name": null
                            }
                          }
                        }
                      ]
                    }
                  ],
                  "notes": ["string"],
                  "recorded": true,
                  "out_of_office": false
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid call ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "description": "Status code"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error message"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error name"
                    }
                  }
                },
                "example": {
                  "statusCode": 400,
                  "message": [
                    "callId must be a number conforming to the specified constraints"
                  ],
                  "error": "Bad Request"
                }
              }
            }
          },
          "404": {
            "description": "Call not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "description": "Status code"
                    },
                    "message": {
                      "type": "string",
                      "description": "The error message"
                    },
                    "error": {
                      "type": "string",
                      "description": "The error name"
                    }
                  }
                },
                "example": {
                  "statusCode": 404,
                  "message": "Call with id '12456' doesn't exist on company '100123'",
                  "error": "Not Found"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
...
      "CallTimes": {
        "type": "object",
        "properties": {
          "total_time": {
            "type": "number",
            "nullable": true,
            "description": "Total time spent in a call"
          },
          "talking_time": {
            "type": "number",
            "nullable": true,
            "description": "Time spent talking to agent/customer"
          },
          "wrap_up_time": {
            "type": "number",
            "nullable": true,
            "description": "Time spent in wrap-up screen"
          },
          "ringing_time": {
            "type": "number",
            "nullable": true,
            "description": "Time spent ringing on an agent"
          },
          "waiting_time": {
            "type": "number",
            "nullable": true,
            "description": "Time from after leaving IVR to getting answered/missed by agent"
          },
          "holding_time": {
            "type": "number",
            "nullable": true,
            "description": "Time spent on hold during a call"
          }
        }
      },
      "ResolvedByCall": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of a call that resolved this call"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call was resolved"
          },
          "agent": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "nullable": true,
                "description": "ID of an agent that resolved this call"
              },
              "name": {
                "type": "string",
                "nullable": true,
                "description": "Name of an agent that resolved this call"
              }
            }
          }
        }
      },
      "AgentStep": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["agent"]
          },
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of an agent"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Name of an agent"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the agent answered/missed the call"
          },
          "call_id": {
            "type": "number",
            "description": "ID of a call that was answered/missed by an agent"
          },
          "call_times": {
            "$ref": "#/components/schemas/CallTimes"
          },
          "group_ids": {
            "type": "array",
            "items": {
              "type": "number"
            },
            "description": "IDs of groups that the agent belongs to"
          },
          "status": {
            "type": "string",
            "enum": ["answered", "missed", "cold_transfer"],
            "description": "Status of the agent call step"
          },
          "reason": {
            "type": "string",
            "enum": ["busy", "not_picked_up", "offline", "idle", "canceled"],
            "description": "Reason why agent missed the call"
          },
          "resolved_by": {
            "type": "number",
            "nullable": true,
            "description": "ID of a call that resolved this call"
          },
          "resolved_by_call": {
            "$ref": "#/components/schemas/ResolvedByCall"
          }
        }
      },
      "QueueStep": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["queue"]
          },
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of a queue"
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Name of a queue"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call was calling the queue"
          },
          "call_times": {
            "$ref": "#/components/schemas/CallTimes"
          },
          "status": {
            "type": "string",
            "enum": ["answered", "missed", "cold_transfer"],
            "description": "Status of the queue call step"
          },
          "reason": {
            "type": "string",
            "enum": [
              "no_agents_online",
              "no_agents_available",
              "no_agents_picked_up",
              "canceled"
            ],
            "description": "Reason why the queue missed the call"
          },
          "resolved_by": {
            "type": "number",
            "nullable": true,
            "description": "ID of a call that resolved this call"
          },
          "resolved_by_call": {
            "$ref": "#/components/schemas/ResolvedByCall"
          },
          "strategy": {
            "type": "string",
            "enum": ["ringall", "rrmemory", "random"],
            "description": "Strategy used by the queue"
          },
          "agent_calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentStep"
            }
          }
        }
      },
      "IVRStep": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["ivr"]
          },
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of an IVR"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call was on the IVR step"
          },
          "total_time": {
            "type": "number",
            "nullable": true,
            "description": "Total time spent in IVR"
          },
          "option": {
            "type": "string",
            "nullable": true,
            "description": "Option pressed by the caller"
          }
        }
      },
      "PlaybackStep": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["playback"]
          },
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of a playback"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call was on the playback step"
          },
          "file": {
            "type": "string",
            "nullable": true,
            "description": "File played"
          },
          "total_time": {
            "type": "number",
            "nullable": true,
            "description": "Total time spent in playback"
          }
        }
      },
      "TransferStep": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["transfer"]
          },
          "id": {
            "type": "number",
            "nullable": true,
            "description": "ID of a transfer"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call was on the transfer step"
          },
          "endpoint": {
            "type": "string",
            "nullable": true,
            "description": "Endpoint of the transfer"
          },
          "status": {
            "type": "string",
            "enum": ["answered", "missed"],
            "description": "Status of the transfer call step"
          }
        }
      },
      "Call": {
        "type": "object",
        "properties": {
          "cdr_id": {
            "type": "number"
          },
          "uuid": {
            "type": "string"
          },
          "company_id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the call started"
          },
          "contact": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "nullable": true,
                "description": "ID of a contact"
              },
              "name": {
                "type": "string",
                "nullable": true,
                "description": "Name of a contact"
              },
              "country": {
                "type": "string",
                "nullable": true,
                "description": "Country of a contact"
              },
              "number": {
                "type": "string",
                "description": "Number of a contact"
              }
            }
          },
          "call_tags": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "nullable": true,
                  "description": "ID of a tag"
                },
                "label": {
                  "type": "string",
                  "nullable": true,
                  "description": "Label of a tag"
                }
              }
            }
          },
          "call_rating": {
            "type": "number",
            "nullable": true,
            "description": "Rating of a call"
          },
          "internal_number": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "nullable": true,
                "description": "ID of an internal number"
              },
              "name": {
                "type": "string",
                "nullable": true,
                "description": "Name of an internal number"
              },
              "number": {
                "type": "string",
                "description": "Phone number of an internal number"
              }
            }
          },
          "call_times": {
            "$ref": "#/components/schemas/CallTimes"
          },
          "direction": {
            "type": "string",
            "enum": ["incoming", "outgoing", "internal"]
          },
          "type": {
            "type": "string",
            "enum": [
              "power_dialer",
              "predictive_dialer",
              "callback",
              "redirected",
              "transferred",
              "regular"
            ]
          },
          "status": {
            "type": "string",
            "enum": ["answered", "missed"]
          },
          "call_steps": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QueueStep"
                },
                {
                  "$ref": "#/components/schemas/AgentStep"
                },
                {
                  "$ref": "#/components/schemas/IVRStep"
                },
                {
                  "$ref": "#/components/schemas/PlaybackStep"
                },
                {
                  "$ref": "#/components/schemas/TransferStep"
                }
              ]
            },
            "description": "List of call steps"
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recorded": {
            "type": "boolean",
            "description": "Indicates if the call was recorded"
          },
          "out_of_office": {
            "type": "boolean",
            "description": "Indicates if the call was made in the out of office hours"
          }
        }
      }
    },
...
  }
}